Simplify No Runtime runtime to load app as slib
Instead of fork/exec'ing an executable, the nort runtime loads a shared
library (using `dlopen`) which is exepcted to contain an `init` and
`handle` function. The runtime is responsible for VSock communication
(as other runtimes are) and defers only JSON serializing/deserializing
to the app.
