+++ appfs/nort/hello/Makefile
@@ -3,7 +3,7 @@ mkfile_path := $(dir $(abspath $(lastword $(MAKEFILE_LIST))))
 all: output.ext2
 
 workload: workload.c
-	echo "apk add alpine-sdk linux-headers; gcc -O3 -o /runtime/workload /runtime/workload.c" | docker run -i --rm -v $(mkfile_path):/runtime alpine
+	echo "apk add alpine-sdk linux-headers; gcc -O3 -o /runtime/workload -fPIC -shared /runtime/workload.c" | docker run -i --rm -v $(mkfile_path):/runtime alpine
 
 output.ext2: workload
 	rm -rf output.ext2 /tmp/lorem.out/
