+++ separate/runtimes/nort/workload.c
@@ -5,25 +5,73 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+#include <dlfcn.h>
 #include <fcntl.h>
 #include <sched.h>
 #include <unistd.h>
 
+#include <arpa/inet.h>
+
 #include <sys/io.h>
 #include <sys/ioctl.h>
 #include <sys/mount.h>
-#include <sys/wait.h>
 
 #include <linux/random.h>
+#include <linux/vm_sockets.h>
 
-#ifndef FAKE_SNAP
+typedef void initfunc(void);
+typedef void responder(char*, uint32_t);
+typedef void handlefunc(char*,uint32_t,responder);
+
+#ifdef TEST
+static int vsock_connect(int cid, int port)
+{
+  return 0;
+}
+#else
+static int vsock_connect(int cid, int port)
+{
+	int fd;
+	struct sockaddr_vm sa = {
+    .svm_family =  AF_VSOCK,
+  };
+	sa.svm_cid = cid;
+	sa.svm_port = port;
+
+	fd = socket(AF_VSOCK, SOCK_STREAM, 0);
+	if (fd < 0) {
+		perror("socket");
+		return -1;
+	}
+
+	if (connect(fd, (struct sockaddr*)&sa, sizeof(sa)) != 0) {
+		perror("connect");
+		close(fd);
+		return -1;
+	}
+
+	return fd;
+}
+#endif
+
+#ifdef TEST
+// Fake snapshot function for local testing
+void snapshot() {
+  setbuf(stdout, NULL);
+  printf("Snapshotting...");
+  for (int i = 0; i < 100000000; i++) {
+    asm("nop");
+  }
+  printf("Done\n");
+}
+#else
 void snapshot() {
   // Make sure we are allowed to perform `outl`
   if (iopl(3)) {perror("iopl"); exit(1);}
 
+  cpu_set_t cset;
   // Let VMM know each of the CPUS is ready for a snapshot
   for (int cpu = 1; ; cpu++) {
-    cpu_set_t cset;
     CPU_ZERO(&cset);
     CPU_SET(cpu, &cset);
     if (sched_setaffinity(0, sizeof(cpu_set_t), &cset) < 0) {
@@ -35,75 +83,53 @@ void snapshot() {
   }
 
   // Finally, signal the VMM to start snapshotting from the main CPU
-  cpu_set_t cset;
   CPU_ZERO(&cset);
   CPU_SET(0, &cset);
   sched_setaffinity(0, sizeof(cset), &cset);
   outl(124, 0x3f0);
 }
-#else
-// Fake snapshot function for local testing
-void snapshot() {
-  setbuf(stdout, NULL);
-  printf("Snapshotting...");
-  for (int i = 0; i < 100000000; i++) {
-    asm("nop");
-  }
-  printf("Done\n");
-}
 #endif
+
+int vsock;
+
+static void respond(char *response, uint32_t response_len) {
+  uint32_t response_len_buf = htonl(response_len);
+  write(vsock, (char*)&response_len_buf, 4);
+  write(vsock, response, response_len);
+}
 
 int main(int argc, char* argv[]) {
   snapshot();
 
+#ifndef TEST
   // Mount the function filesystem
   if (mount("/dev/vdb", "/srv", "ext2", MS_RDONLY, NULL) < 0) {
     perror("mount");
   }
+#endif
 
-  // We'll signal readiness between parent & child using a pipe.
-  int p2c_pipefds[2];
-  pipe(p2c_pipefds);
-  int c2p_pipefds[2];
-  pipe(c2p_pipefds);
+  void *libhandle = dlopen(argv[1], RTLD_NOW);
+  initfunc *init = dlsym(libhandle, "init");
+  handlefunc *handle = dlsym(libhandle, "handle");
 
-  int pid = fork();
-  if (pid == 0) {
-    // Child process
+  init();
 
-    // Close parent's side of pipes
-    close(p2c_pipefds[1]);
-    close(c2p_pipefds[0]);
+  snapshot();
 
-    dup2(p2c_pipefds[0], 3); // read from parent
-    dup2(c2p_pipefds[1], 4); // write to parent
-    // Run, child!
-    char *argv2[2] = { argv[1], NULL };
-    exit(execv(argv[1], argv2));
-  } else {
-    // Parent process
+  vsock = vsock_connect(2, 1234);
+  if (vsock < 0) {
+    exit(-1);
+  }
+  printf("connected\n");
 
-    // Close child's side of pipes
-    close(p2c_pipefds[0]);
-    close(c2p_pipefds[1]);
+  for (;;) {
+    uint32_t buffer_length = 0;
+    read(vsock, (char*)&buffer_length, 4);
+    buffer_length = ntohl(buffer_length);
+    char *buffer = (char*)malloc(buffer_length);
+    read(vsock, buffer, buffer_length);
 
-    // Wait for child pipe to close;
-    char throwaway[1];
-    read(c2p_pipefds[0], throwaway, 1);
-    close(c2p_pipefds[0]);
-
-    snapshot();
-
-    // Signal child we're done snapshotting and it can open vsock
-    write(p2c_pipefds[1], throwaway, 1);
-    close(p2c_pipefds[1]);
-
-    // Wait for the child to exit
-    int status;
-    do {
-      waitpid(pid, &status, 0);
-    } while (!WIFEXITED(status));
-    exit(WEXITSTATUS(status));
+    handle(buffer, buffer_length, &respond);
   }
 }
